feat: now MyTripsItem is clickable
