+++ app/src/main/res/values-en/strings.xml
@@ -17,7 +17,7 @@
     <string name="pick_shake_trevel">Pick, shake and travel!</string>
     <string name="your_travel_guide">Your travel guide when you shake your phone.</string>
     <string name="discover_exciting">Discover exciting destinations \nwith just a gesture.</string>
-    <string name="discover_more_trips">Discover more trips</string>
+    <string name="discover_more_trips">Recommended trips</string>
     <string name="nearly">Nearly %1$s</string>
     <string name="no_results_found">No results found</string>
     <string name="my_trips">My trips</string>
