+++ app/src/main/res/values-es/strings.xml
@@ -18,7 +18,7 @@
     <string name="pick_shake_trevel">¡Elige, agita y viaja!</string>
     <string name="your_travel_guide">Tu guía de viaje al agitar el móvil.</string>
     <string name="discover_exciting">Descubre destinos emocionantes \ncon solo un gesto.</string>
-    <string name="discover_more_trips">Descubre más viajes</string>
+    <string name="discover_more_trips">Viajes recomendados</string>
     <string name="nearly">Cerca de %1$s</string>
     <string name="no_results_found">No se encontraron resultados</string>
     <string name="my_trips">Mis viajes</string>
